Added option to use double space "  " as a delimiter in the checkin data
Fixed a bug where the different message formats were checked even though the newFormat had been used.

Fixed a bug where a false negative on using the newFormat happened.

Allow MTR as an abbreviation for meters

Updated the program to read the credentials filename and spreadsheet ID from the roster.txt file.

Updated the program so that it would read attachments and save the CSV data.