Fixed a broken link to the comments file.
countryLong fixed
countries expanded to cover all countries and populate the variables from the static list

logic to find checkinCountry and checkinCountryLong simplified and covers all countries

CheckinItems needed to be nulled and not filled if the items don't exist properly.

Fixed handling checkin data if the beginning only has a single # character but is otherwise correct for the new format.